Remove enable/disablefrom non-measuring routines and tweaked time delay in get_full luminostiy routine.
It's not necessary to enable the oscillator and start measuring when changing the gain or itegration time.  You can read/write to the registers without doing so.  From a power perspective it's better not to.

Added code to check whether integration complete.  This allowed determination of
minimum delay for measurement.  Reduced from 1.x second to .2+ sec depending on integration time.  The order of reading the registers appears to be significant.  Lux values stabilized after switching order.  Data sheet actually recommends reading it in a block.